% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfit_automatic.R
\name{fitGaussPeakLinBg}
\alias{fitGaussPeakLinBg}
\title{fit peak to gaussian with linear bacground}
\usage{
fitGaussPeakLinBg(x, y, range)
}
\arguments{
\item{x}{x data for fit}

\item{y}{y data for fit}

\item{range}{range of x data that will be used for graph drawing}
}
\value{
list(fit, data)
}
\description{
Fit peak to gaussian function with linear background. Fitting
             interval should be defined by two end points which shoud be
             selected by two clicks at graph. Function will return fit.
             Function will return list with parameters of fit and data.frame
             with data (x,y) in interval used for fit (between two selected
             points). Linear background will be constructed as linear
             function thorought two end points.
}
\examples{
 \dontrun{fitGaussPeakLinBg(x, y, range)}
}
\author{
Krunoslav Juraic
}

