% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/peak_profile_functions.R
\name{gaussianArea}
\alias{gaussianArea}
\title{Gaussian area}
\usage{
gaussianArea(x, p)
}
\arguments{
\item{x}{x value}

\item{p}{peak paramteres (p[1] = amplitude, p[2] = center, p[3] = width st.dev.)}
}
\value{
y y value
}
\description{
Gaussian area function
}
\examples{
gaussianArea(x = 0.1, p = c(1, 0, .5))
}
\author{
K. Juraic
}

