% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_profile_functions.R
\name{gaussianAmpl}
\alias{gaussianAmpl}
\title{Gaussian amplitud}
\usage{
gaussianAmpl(x, y0, A, xc, w)
}
\arguments{
\item{x}{x value}

\item{y0}{backgroubd}

\item{A}{amplitude}

\item{xc}{peak center}

\item{w}{peak width}
}
\value{
y y value
}
\description{
Gaussian amplitude function
}
\examples{
   \dontrun{gaussianAmpl(x = 0.1, y0 = 0, A = 1, xc = 0, w = 1)}
}
\references{
PeakFit v4.11
}
\author{
K. Juraic
}
