% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_profile_functions.R
\name{gaussianArea}
\alias{gaussianArea}
\title{Gaussian area}
\usage{
gaussianArea(x, y0, A, xc, w)
}
\arguments{
\item{x}{x value}

\item{y0}{backgroubd}

\item{A}{amplitude}

\item{xc}{peak center}

\item{w}{peak width}
}
\value{
y y value
}
\description{
Gaussian area function
}
\examples{
   \dontrun{gaussianArea(x = 0.1, y0 = 0, A = 1, xc = 0, w = 1)}
}
\author{
K. Juraic
}

